pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "futures-bot"
version = "0.1.0"
description = "Bybit futures auto-trading bot with self-improvement loop"
requires-python = ">=3.12"
dependencies = [
    "pybit>=5.9.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "ta-lib>=0.5.1",
    "scikit-learn>=1.5.0",
    "xgboost>=2.1.0",
    "torch>=2.4.0",
    "psycopg2-binary>=2.9.9",
    "sqlalchemy>=2.0.0",
    "apscheduler>=3.10.0",
    "python-telegram-bot>=21.0",
    "python-dotenv>=1.0.0",
    "vectorbt>=0.26.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
